Separate file for mocked metrics
